+++ plugin/src/main/kotlin/com/testknight/actions/testlist/LoadTestAction.kt
@@ -1,4 +1,4 @@
-package com.testknight.actions.testcases
+package com.testknight.actions.testlist
 
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
