+++ plugin/src/main/kotlin/com/testknight/utilities/UserInterfaceHelper.kt
@@ -9,8 +9,8 @@ import com.intellij.psi.PsiManager
 import com.intellij.ui.components.JBPanelWithEmptyText
 import com.intellij.ui.components.JBScrollPane
 import com.intellij.ui.content.impl.ContentImpl
-import com.testknight.actions.testcases.ClearTestAction
-import com.testknight.actions.testcases.LoadTestAction
+import com.testknight.actions.testlist.ClearTestAction
+import com.testknight.actions.testlist.LoadTestAction
 import javax.swing.JTabbedPane
 import javax.swing.JViewport
