+++ plugin/src/main/kotlin/com/testknight/views/UserInterface.kt
@@ -14,23 +14,25 @@ import com.intellij.ui.components.JBScrollPane
 import com.intellij.ui.components.JBTabbedPane
 import com.intellij.ui.table.JBTable
 import com.intellij.ui.treeStructure.Tree
-import com.testknight.actions.AddItemChecklistAction
-import com.testknight.actions.DeleteElementChecklistAction
-import com.testknight.actions.EditItemChecklistAction
+import com.testknight.actions.checklist.AddItemChecklistAction
+import com.testknight.actions.checklist.DeleteElementChecklistAction
+import com.testknight.actions.checklist.EditItemChecklistAction
 import com.testknight.actions.GenerateTestMethodAction
 import com.testknight.actions.ShowCoverageDiffAction
-import com.testknight.actions.testcases.TestListTraceabilityAction
-import com.testknight.listeners.CheckListKeyboardListener
-import com.testknight.listeners.CheckedNodeListener
-import com.testknight.listeners.ChecklistMouseListener
-import com.testknight.listeners.ChecklistSelectionListener
+import com.testknight.actions.testlist.TestListTraceabilityAction
+import com.testknight.listeners.checklist.CheckListKeyboardListener
+import com.testknight.listeners.checklist.CheckedNodeListener
+import com.testknight.listeners.checklist.ChecklistMouseListener
+import com.testknight.listeners.checklist.ChecklistSelectionListener
 import com.testknight.listeners.PsiTreeListener
-import com.testknight.listeners.TestListKeyboardListener
-import com.testknight.listeners.TestListMouseListener
-import com.testknight.listeners.TestListSelectionListener
-import com.testknight.services.ChecklistTreeService
+import com.testknight.listeners.testlist.TestListKeyboardListener
+import com.testknight.listeners.testlist.TestListMouseListener
+import com.testknight.listeners.testlist.TestListSelectionListener
+import com.testknight.services.checklist.ChecklistTreeService
 import com.testknight.services.LoadTestsService
 import com.testknight.utilities.UserInterfaceHelper
+import com.testknight.views.coverage.ButtonColumn
+import com.testknight.views.coverage.CoverageStatsCellRenderer
 import com.testknight.views.trees.ChecklistCellEditor
 import com.testknight.views.trees.TestListCellRenderer
 import org.jetbrains.annotations.NotNull
